Fix remaining warnings
Former-commit-id: d6131bbee5506c1d5fc060d95b246278f48d0627
Former-commit-id: 529feb749e9f4100edd3ac5425fa773b7123d7a4
Former-commit-id: c5615937be7fbbe6ac6eab70d77c29850e14477f
Former-commit-id: 67ce6e6b72a8cea3343614754d2d9c6c299e5345 [formerly e7bd2043c1b2d83c0432b32ea2de86f155a237df]
Former-commit-id: 0d505125e0b19a287371984408240866f0efda66

+++ src/tamm/memory_manager_ga.hpp
@@ -105,7 +105,8 @@ class MemoryManagerGA : public MemoryManager {
       pmr->map_[0] = 0; // @note this is not set by MPI_Exscan
       std::partial_sum(pmr->map_.begin(), pmr->map_.begin()+nranks, pmr->map_.begin());
       std::string array_name{"array_name"};
-      for(block = nranks; block>0 && pmr->map_[block-1] == dim; --block) {
+      
+      for(block = nranks; block>0 && static_cast<int64_t>(pmr->map_[block-1]) == dim; --block) {
         //no-op
       }
 
@@ -122,8 +123,8 @@ class MemoryManagerGA : public MemoryManager {
 
     int64_t lo, hi;//, ld;
     NGA_Distribution64(pmr->ga_, pg_.rank().value(), &lo, &hi);
-    EXPECTS(nels<=0 || lo == pmr->map_[pg_.rank().value()]);
-    EXPECTS(nels<=0 || hi == pmr->map_[pg_.rank().value()] + nels - 1);
+    EXPECTS(nels<=0 || lo == static_cast<int64_t>(pmr->map_[pg_.rank().value()]));
+    EXPECTS(nels<=0 || hi == static_cast<int64_t>(pmr->map_[pg_.rank().value()]) + nels - 1);
     pmr->set_status(AllocationStatus::created);
     return pmr;
   }
